// packages/router/src/models.d.ts
export type QueryParamsHandling = 'merge' | 'preserve' | 'replace' | '';